LPD-77985 [BE] Move logic to ObjectFieldBusinessType
