+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/dto/v1_0/converter/ObjectEntryDTOConverter.java
@@ -8,19 +8,9 @@ package com.liferay.object.rest.internal.dto.v1_0.converter;
 import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.service.AssetCategoryLocalService;
 import com.liferay.asset.kernel.service.AssetTagLocalService;
-import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.document.library.kernel.model.DLFolder;
-import com.liferay.document.library.kernel.processor.PDFProcessorUtil;
-import com.liferay.document.library.kernel.service.DLAppLocalService;
-import com.liferay.document.library.kernel.service.DLAppService;
-import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
-import com.liferay.document.library.util.DLURLHelper;
-import com.liferay.exportimport.attachment.ExportImportAttachmentManager;
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.headless.delivery.dto.v1_0.Comment;
 import com.liferay.headless.delivery.dto.v1_0.util.CommentUtil;
-import com.liferay.list.type.model.ListTypeEntry;
-import com.liferay.list.type.service.ListTypeEntryLocalService;
 import com.liferay.object.constants.ObjectActionKeys;
 import com.liferay.object.constants.ObjectDefinitionConstants;
 import com.liferay.object.constants.ObjectFieldConstants;
@@ -28,6 +18,8 @@ import com.liferay.object.constants.ObjectFieldSettingConstants;
 import com.liferay.object.constants.ObjectRelationshipConstants;
 import com.liferay.object.entry.util.ObjectEntryDTOConverterUtil;
 import com.liferay.object.entry.util.ObjectEntryValuesUtil;
+import com.liferay.object.field.business.type.ObjectFieldBusinessType;
+import com.liferay.object.field.business.type.ObjectFieldBusinessTypeRegistry;
 import com.liferay.object.field.setting.util.ObjectFieldSettingUtil;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntryFolder;
@@ -38,12 +30,8 @@ import com.liferay.object.model.ObjectField;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.related.models.ObjectRelatedModelsProvider;
 import com.liferay.object.related.models.ObjectRelatedModelsProviderRegistry;
-import com.liferay.object.rest.dto.v1_0.Assignee;
 import com.liferay.object.rest.dto.v1_0.AuditEvent;
 import com.liferay.object.rest.dto.v1_0.AuditFieldChange;
-import com.liferay.object.rest.dto.v1_0.FileEntry;
-import com.liferay.object.rest.dto.v1_0.Folder;
-import com.liferay.object.rest.dto.v1_0.ListEntry;
 import com.liferay.object.rest.dto.v1_0.ObjectDefinitionBrief;
 import com.liferay.object.rest.dto.v1_0.ObjectEntry;
 import com.liferay.object.rest.dto.v1_0.Status;
@@ -51,7 +39,7 @@ import com.liferay.object.rest.dto.v1_0.SystemProperties;
 import com.liferay.object.rest.dto.v1_0.TaxonomyCategoryBrief;
 import com.liferay.object.rest.dto.v1_0.Version;
 import com.liferay.object.rest.dto.v1_0.util.CreatorUtil;
-import com.liferay.object.rest.dto.v1_0.util.LinkUtil;
+import com.liferay.object.rest.dto.v1_0.util.ListEntryUtil;
 import com.liferay.object.rest.internal.dto.v1_0.util.TaxonomyCategoryBriefUtil;
 import com.liferay.object.scope.ObjectScopeProvider;
 import com.liferay.object.scope.ObjectScopeProviderRegistry;
@@ -70,7 +58,6 @@ import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.comment.CommentManager;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
@@ -81,10 +68,7 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.GroupConstants;
-import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.repository.model.FileVersion;
 import com.liferay.portal.kernel.security.auth.GuestOrUserUtil;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
@@ -92,18 +76,11 @@ import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.PermissionService;
 import com.liferay.portal.kernel.service.ResourceActionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Base64;
-import com.liferay.portal.kernel.util.DateUtil;
-import com.liferay.portal.kernel.util.File;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.HtmlParserUtil;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortalUtil;
@@ -111,7 +88,6 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.language.LanguageResources;
-import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
 import com.liferay.portal.security.audit.event.generators.constants.EventTypes;
 import com.liferay.portal.security.audit.storage.service.AuditEventLocalService;
 import com.liferay.portal.vulcan.dto.converter.DTOConverter;
@@ -127,7 +103,6 @@ import com.liferay.portal.vulcan.jaxrs.extension.ExtendedEntity;
 import com.liferay.portal.vulcan.permission.Permission;
 import com.liferay.portal.vulcan.permission.PermissionUtil;
 import com.liferay.portal.vulcan.scope.Scope;
-import com.liferay.portal.vulcan.util.LocalizedMapUtil;
 import com.liferay.trash.model.TrashEntry;
 import com.liferay.trash.service.TrashEntryLocalService;
 
@@ -135,12 +110,7 @@ import jakarta.ws.rs.core.UriInfo;
 
 import java.io.Serializable;
 
-import java.sql.Timestamp;
-
-import java.text.SimpleDateFormat;
-
 import java.util.Collection;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -240,7 +210,7 @@ public class ObjectEntryDTOConverter
 
 			properties.put(
 				objectField.getName(),
-				_getListEntry(
+				ListEntryUtil.toListEntry(
 					dtoConverterContext, map.get("key"),
 					objectField.getListTypeDefinitionId()));
 
@@ -673,63 +643,6 @@ public class ObjectEntryDTOConverter
 		}
 	}
 
-	private Assignee _getAssignee(Map<String, Serializable> assigneeMap) {
-		if (assigneeMap.containsKey("externalReferenceCode")) {
-			return Assignee.toDTO(_jsonFactory.looseSerializeDeep(assigneeMap));
-		}
-
-		String className = _portal.fetchClassName(
-			MapUtil.getLong(assigneeMap, "classNameId"));
-
-		if (StringUtil.equals(className, Role.class.getName())) {
-			Role role = _roleLocalService.fetchRole(
-				MapUtil.getLong(assigneeMap, "classPK"));
-
-			if (role == null) {
-				return new Assignee();
-			}
-
-			return new Assignee() {
-				{
-					setExternalReferenceCode(role::getExternalReferenceCode);
-					setName(role::getName);
-					setType(() -> Type.ROLE);
-				}
-			};
-		}
-		else if (StringUtil.equals(className, User.class.getName())) {
-			User user = _userLocalService.fetchUser(
-				MapUtil.getLong(assigneeMap, "classPK"));
-
-			if (user == null) {
-				return new Assignee();
-			}
-
-			return new Assignee() {
-				{
-					setExternalReferenceCode(user::getExternalReferenceCode);
-					setName(user::getFullName);
-					setPortrait(
-						() -> {
-							if (user.getPortraitId() == 0) {
-								return null;
-							}
-
-							return user.getPortraitURL(
-								new ThemeDisplay() {
-									{
-										setPathImage(_portal.getPathImage());
-									}
-								});
-						});
-					setType(() -> Type.USER);
-				}
-			};
-		}
-
-		return new Assignee();
-	}
-
 	private <T> T _getAttribute(
 		ObjectEntryVersion objectEntryVersion,
 		Function<ObjectEntryVersion, T> objectEntryVersionGetterFunction,
@@ -743,26 +656,6 @@ public class ObjectEntryDTOConverter
 
 		return serviceBuilderObjectEntryGetterFunction.apply(
 			serviceBuilderObjectEntry);
-	}
-
-	private String _getDateString(
-		ObjectField objectField, Timestamp timestamp) {
-
-		String pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS";
-
-		if (objectField.compareBusinessType(
-				ObjectFieldConstants.BUSINESS_TYPE_DATE) ||
-			StringUtil.equals(
-				ObjectFieldSettingUtil.getValue(
-					ObjectFieldSettingConstants.NAME_TIME_STORAGE, objectField),
-				ObjectFieldSettingConstants.VALUE_CONVERT_TO_UTC)) {
-
-			pattern += "'Z'";
-		}
-
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat(pattern);
-
-		return simpleDateFormat.format(timestamp);
 	}
 
 	private DTOConverterContext _getDTOConverterContext(
@@ -789,195 +682,6 @@ public class ObjectEntryDTOConverter
 		}
 
 		return defaultDTOConverterContext;
-	}
-
-	private FileEntry _getFileEntry(
-			ObjectDefinition objectDefinition,
-			com.liferay.object.model.ObjectEntry objectEntry,
-			ObjectField objectField, long fileEntryId, String objectFieldName)
-		throws Exception {
-
-		FileEntry fileEntry = new FileEntry();
-
-		DLFileEntry dlFileEntry = _dLFileEntryLocalService.fetchDLFileEntry(
-			fileEntryId);
-
-		if (dlFileEntry == null) {
-			return fileEntry;
-		}
-
-		LiferayFileEntry liferayFileEntry = new LiferayFileEntry(dlFileEntry);
-
-		FileVersion fileVersion = liferayFileEntry.getFileVersion();
-
-		fileEntry.setAlternativeText(fileVersion::getDescription);
-
-		fileEntry.setExternalReferenceCode(
-			dlFileEntry::getExternalReferenceCode);
-
-		fileEntry.setFileBase64(
-			() -> (String)NestedFieldsSupplier.supply(
-				objectFieldName + ".fileBase64",
-				fieldName -> Base64.encode(
-					_file.getBytes(dlFileEntry.getContentStream()))));
-		fileEntry.setFileURL(
-			() -> {
-				if (!Objects.equals(
-						ObjectFieldSettingConstants.VALUE_USER_COMPUTER,
-						ObjectFieldSettingUtil.getValue(
-							ObjectFieldSettingConstants.NAME_FILE_SOURCE,
-							objectField)) ||
-					GetterUtil.getBoolean(
-						ObjectFieldSettingUtil.getValue(
-							ObjectFieldSettingConstants.
-								NAME_SHOW_FILES_IN_DOCS_AND_MEDIA,
-							objectField.getObjectFieldSettings()))) {
-
-					return null;
-				}
-
-				return _exportImportAttachmentManager.getFileURL(dlFileEntry);
-			});
-		fileEntry.setFolder(
-			() -> (Folder)NestedFieldsSupplier.supply(
-				objectFieldName + ".folder",
-				fieldName -> {
-					if (!Objects.equals(
-							ObjectFieldSettingConstants.VALUE_DOCS_AND_MEDIA,
-							ObjectFieldSettingUtil.getValue(
-								ObjectFieldSettingConstants.NAME_FILE_SOURCE,
-								objectField))) {
-
-						return null;
-					}
-
-					Folder folder = new Folder();
-
-					folder.setExternalReferenceCode(
-						() -> {
-							if (dlFileEntry.getFolderId() == 0) {
-								return null;
-							}
-
-							DLFolder dlFolder = dlFileEntry.getFolder();
-
-							return dlFolder.getExternalReferenceCode();
-						});
-					folder.setSiteId(dlFileEntry::getGroupId);
-
-					return folder;
-				}));
-
-		fileEntry.setId(dlFileEntry::getFileEntryId);
-		fileEntry.setLink(
-			() -> LinkUtil.toLink(
-				_dlAppService, dlFileEntry, _dlURLHelper,
-				objectEntry.getGroupId(),
-				objectDefinition.getExternalReferenceCode(), objectEntry,
-				_objectEntryService, objectField,
-				GuestOrUserUtil.getPermissionChecker(), _portal));
-
-		fileEntry.setMetadata(
-			() -> NestedFieldsSupplier.supply(
-				objectFieldName + ".metadata",
-				fieldName -> {
-					if ((fileVersion == null) || (fileVersion.getSize() == 0) ||
-						(!PDFProcessorUtil.hasImages(fileVersion) &&
-						 !PDFProcessorUtil.isDocumentSupported(
-							 fileVersion.getMimeType()))) {
-
-						return null;
-					}
-
-					return HashMapBuilder.<String, Object>put(
-						"numberOfPages",
-						PDFProcessorUtil.getPreviewFileCount(fileVersion)
-					).build();
-				}));
-		fileEntry.setMimeType(dlFileEntry::getMimeType);
-		fileEntry.setName(dlFileEntry::getFileName);
-		fileEntry.setPreviewURL(
-			() -> NestedFieldsSupplier.supply(
-				objectFieldName + ".previewURL",
-				fieldName -> {
-					String previewURL = _getPreviewURL(liferayFileEntry);
-
-					if (Validator.isNull(previewURL)) {
-						return null;
-					}
-
-					return previewURL;
-				}));
-		fileEntry.setScope(
-			() -> {
-				if ((objectEntry.getGroupId() == dlFileEntry.getGroupId()) &&
-					!Objects.equals(
-						objectDefinition.getScope(),
-						ObjectDefinitionConstants.SCOPE_COMPANY)) {
-
-					return null;
-				}
-
-				Group group = _groupLocalService.getGroup(
-					dlFileEntry.getGroupId());
-
-				Scope.Type type =
-					(group.getType() == GroupConstants.TYPE_DEPOT) ?
-						Scope.Type.ASSET_LIBRARY : Scope.Type.SITE;
-
-				return Scope.ofReference(
-					group.getExternalReferenceCode(), type);
-			});
-		fileEntry.setThumbnailURL(
-			() -> NestedFieldsSupplier.supply(
-				objectFieldName + ".thumbnailURL",
-				fieldName -> {
-					String thumbnailURL = _dlURLHelper.getThumbnailSrc(
-						new LiferayFileEntry(dlFileEntry), null);
-
-					if (Validator.isNull(thumbnailURL)) {
-						return null;
-					}
-
-					return thumbnailURL;
-				}));
-
-		return fileEntry;
-	}
-
-	private ListEntry _getListEntry(
-		DTOConverterContext dtoConverterContext, String key,
-		long listTypeDefinitionId) {
-
-		if (StringUtil.equals(key, StringPool.BLANK)) {
-			return new ListEntry() {
-				{
-					setKey(() -> StringPool.BLANK);
-					setName(() -> StringPool.BLANK);
-				}
-			};
-		}
-
-		ListTypeEntry listTypeEntry =
-			_listTypeEntryLocalService.fetchListTypeEntry(
-				listTypeDefinitionId, key);
-
-		if (listTypeEntry == null) {
-			return null;
-		}
-
-		return new ListEntry() {
-			{
-				setKey(listTypeEntry::getKey);
-				setName(
-					() -> listTypeEntry.getName(
-						dtoConverterContext.getLocale()));
-				setName_i18n(
-					() -> LocalizedMapUtil.getI18nMap(
-						dtoConverterContext.isAcceptAllLanguages(),
-						listTypeEntry.getNameMap()));
-			}
-		};
 	}
 
 	private Serializable _getLocalizedValue(
@@ -1126,20 +830,6 @@ public class ObjectEntryDTOConverter
 		return objectDefinition;
 	}
 
-	private String _getPreviewURL(LiferayFileEntry liferayFileEntry)
-		throws PortalException {
-
-		if (StringUtil.startsWith(liferayFileEntry.getMimeType(), "video/")) {
-			return _dlURLHelper.getPreviewURL(
-				liferayFileEntry, liferayFileEntry.getFileVersion(), null,
-				"&videoEmbed=true", true, true);
-		}
-
-		return _dlURLHelper.getPreviewURL(
-			liferayFileEntry, liferayFileEntry.getFileVersion(), null,
-			StringPool.BLANK, false, false);
-	}
-
 	private String _getScopeKey(
 		ObjectDefinition objectDefinition,
 		com.liferay.object.model.ObjectEntry objectEntry) {
@@ -1168,134 +858,13 @@ public class ObjectEntryDTOConverter
 			ObjectField objectField, Serializable serializable)
 		throws Exception {
 
-		if (objectField.compareBusinessType(
-				ObjectFieldConstants.BUSINESS_TYPE_ASSIGNEE)) {
+		ObjectFieldBusinessType objectFieldBusinessType =
+			_objectFieldBusinessTypeRegistry.getObjectFieldBusinessType(
+				objectField.getBusinessType());
 
-			if (serializable instanceof Assignee) {
-				return serializable;
-			}
-			else if (serializable instanceof Map) {
-				return _getAssignee((Map<String, Serializable>)serializable);
-			}
-
-			return null;
-		}
-		else if (objectField.compareBusinessType(
-					ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
-
-			if (serializable instanceof FileEntry) {
-				return serializable;
-			}
-
-			long fileEntryId = 0;
-
-			if (serializable instanceof Long) {
-				fileEntryId = GetterUtil.getLong(serializable);
-			}
-			else if (serializable instanceof Map) {
-				Map<String, Serializable> map =
-					(Map<String, Serializable>)serializable;
-
-				fileEntryId = GetterUtil.getLong(map.get("id"));
-			}
-
-			if (fileEntryId == 0) {
-				return null;
-			}
-
-			return _getFileEntry(
-				objectDefinition, objectEntry, objectField, fileEntryId,
-				objectField.getName());
-		}
-		else if (objectField.compareBusinessType(
-					ObjectFieldConstants.BUSINESS_TYPE_DATE) ||
-				 objectField.compareBusinessType(
-					 ObjectFieldConstants.BUSINESS_TYPE_DATE_TIME)) {
-
-			if (Validator.isNull(serializable)) {
-				return null;
-			}
-
-			if (serializable instanceof String) {
-				Date date = DateUtil.parseDate(
-					"yyyy-MM-dd", (String)serializable,
-					LocaleUtil.getSiteDefault());
-
-				serializable = new Timestamp(date.getTime());
-			}
-
-			Timestamp timestamp = (Timestamp)serializable;
-
-			if (timestamp == null) {
-				return null;
-			}
-
-			return _getDateString(objectField, timestamp);
-		}
-		else if (objectField.compareBusinessType(
-					ObjectFieldConstants.BUSINESS_TYPE_MULTISELECT_PICKLIST)) {
-
-			if (objectField.getListTypeDefinitionId() == 0) {
-				return null;
-			}
-
-			if (serializable instanceof List) {
-				return serializable;
-			}
-
-			String[] keys = null;
-
-			if (serializable instanceof Object[]) {
-				keys = TransformUtil.transform(
-					(Object[])serializable,
-					object -> {
-						if (!(object instanceof Map)) {
-							return null;
-						}
-
-						return MapUtil.getString(
-							(Map<String, String>)object, "key");
-					},
-					String.class);
-			}
-			else if (serializable instanceof String) {
-				keys = StringUtil.split(
-					(String)serializable, StringPool.COMMA_AND_SPACE);
-			}
-
-			return (Serializable)TransformUtil.transformToList(
-				keys,
-				key -> _getListEntry(
-					dtoConverterContext, key,
-					objectField.getListTypeDefinitionId()));
-		}
-		else if (objectField.compareBusinessType(
-					ObjectFieldConstants.BUSINESS_TYPE_PICKLIST)) {
-
-			if (objectField.getListTypeDefinitionId() == 0) {
-				return null;
-			}
-
-			if (serializable instanceof ListEntry) {
-				return serializable;
-			}
-
-			String key = null;
-
-			if (serializable instanceof Map) {
-				key = MapUtil.getString(
-					(Map<String, String>)serializable, "key");
-			}
-			else if (serializable instanceof String) {
-				key = (String)serializable;
-			}
-
-			return _getListEntry(
-				dtoConverterContext, key,
-				objectField.getListTypeDefinitionId());
-		}
-
-		return serializable;
+		return objectFieldBusinessType.getDTOValue(
+			dtoConverterContext, objectDefinition, objectEntry, objectField,
+			serializable);
 	}
 
 	private boolean _hasRootModelHierarchyNestedField() {
@@ -1743,25 +1312,7 @@ public class ObjectEntryDTOConverter
 	private DiscussionPermission _discussionPermission;
 
 	@Reference
-	private DLAppLocalService _dlAppLocalService;
-
-	@Reference
-	private DLAppService _dlAppService;
-
-	@Reference
-	private DLFileEntryLocalService _dLFileEntryLocalService;
-
-	@Reference
-	private DLURLHelper _dlURLHelper;
-
-	@Reference
-	private ExportImportAttachmentManager _exportImportAttachmentManager;
-
-	@Reference
 	private ExtensionProviderRegistry _extensionProviderRegistry;
-
-	@Reference
-	private File _file;
 
 	@Reference
 	private GroupLocalService _groupLocalService;
@@ -1771,9 +1322,6 @@ public class ObjectEntryDTOConverter
 
 	@Reference
 	private Language _language;
-
-	@Reference
-	private ListTypeEntryLocalService _listTypeEntryLocalService;
 
 	private ObjectDefinition _objectDefinition;
 
@@ -1788,6 +1336,9 @@ public class ObjectEntryDTOConverter
 
 	@Reference
 	private ObjectEntryService _objectEntryService;
+
+	@Reference
+	private ObjectFieldBusinessTypeRegistry _objectFieldBusinessTypeRegistry;
 
 	@Reference
 	private ObjectFieldLocalService _objectFieldLocalService;
@@ -1810,9 +1361,6 @@ public class ObjectEntryDTOConverter
 
 	@Reference
 	private ResourceActionLocalService _resourceActionLocalService;
-
-	@Reference
-	private RoleLocalService _roleLocalService;
 
 	@Reference
 	private SystemObjectDefinitionManagerRegistry
