+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/field/business/type/AttachmentObjectFieldBusinessType.java
@@ -5,29 +5,60 @@
 
 package com.liferay.object.internal.field.business.type;
 
+import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.model.DLFolder;
+import com.liferay.document.library.kernel.processor.PDFProcessorUtil;
+import com.liferay.document.library.kernel.service.DLAppService;
+import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.exportimport.attachment.ExportImportAttachmentManager;
+import com.liferay.object.constants.ObjectDefinitionConstants;
 import com.liferay.object.constants.ObjectFieldConstants;
 import com.liferay.object.constants.ObjectFieldSettingConstants;
 import com.liferay.object.dynamic.data.mapping.form.field.type.constants.ObjectDDMFormFieldTypeConstants;
 import com.liferay.object.exception.ObjectFieldSettingValueException;
 import com.liferay.object.field.business.type.ObjectFieldBusinessType;
 import com.liferay.object.field.render.ObjectFieldRenderingContext;
+import com.liferay.object.field.setting.util.ObjectFieldSettingUtil;
 import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.model.ObjectField;
 import com.liferay.object.model.ObjectFieldSetting;
+import com.liferay.object.rest.dto.v1_0.FileEntry;
+import com.liferay.object.rest.dto.v1_0.Folder;
+import com.liferay.object.rest.dto.v1_0.util.LinkUtil;
 import com.liferay.object.scope.ObjectScopeProvider;
 import com.liferay.object.scope.ObjectScopeProviderRegistry;
 import com.liferay.object.service.ObjectDefinitionLocalService;
+import com.liferay.object.service.ObjectEntryService;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.GroupConstants;
+import com.liferay.portal.kernel.repository.model.FileVersion;
+import com.liferay.portal.kernel.security.auth.GuestOrUserUtil;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.util.Base64;
+import com.liferay.portal.kernel.util.File;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.SetUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
+import com.liferay.portal.vulcan.dto.converter.DTOConverterContext;
 import com.liferay.portal.vulcan.extension.PropertyDefinition;
+import com.liferay.portal.vulcan.fields.NestedFieldsSupplier;
+import com.liferay.portal.vulcan.scope.Scope;
+
+import java.io.Serializable;
 
 import java.math.BigDecimal;
 
@@ -83,6 +114,71 @@ public class AttachmentObjectFieldBusinessType
 		}
 
 		return super.getDisplayContextValue(objectField, userId, values);
+	}
+
+	@Override
+	public Serializable getDTOValue(
+			DTOConverterContext dtoConverterContext,
+			ObjectDefinition objectDefinition, ObjectEntry objectEntry,
+			ObjectField objectField, Serializable serializable)
+		throws Exception {
+
+		if (serializable instanceof FileEntry) {
+			return serializable;
+		}
+
+		long fileEntryId = 0;
+
+		if (serializable instanceof Long) {
+			fileEntryId = GetterUtil.getLong(serializable);
+		}
+		else if (serializable instanceof Map) {
+			fileEntryId = MapUtil.getLong(
+				(Map<String, Serializable>)serializable, "id");
+		}
+
+		if (fileEntryId == 0) {
+			return null;
+		}
+
+		DLFileEntry dlFileEntry = _dLFileEntryLocalService.fetchDLFileEntry(
+			fileEntryId);
+
+		if (dlFileEntry == null) {
+			return new FileEntry();
+		}
+
+		LiferayFileEntry liferayFileEntry = new LiferayFileEntry(dlFileEntry);
+
+		FileVersion fileVersion = liferayFileEntry.getFileVersion();
+
+		return new FileEntry() {
+			{
+				setAlternativeText(fileVersion::getDescription);
+				setExternalReferenceCode(dlFileEntry::getExternalReferenceCode);
+				setFileBase64(() -> _getFileBase64(dlFileEntry, objectField));
+				setFileURL(() -> _getFileURL(dlFileEntry, objectField));
+				setFolder(() -> _getFolder(dlFileEntry, objectField));
+				setId(dlFileEntry::getFileEntryId);
+				setLink(
+					() -> LinkUtil.toLink(
+						_dlAppService, dlFileEntry, _dlURLHelper,
+						objectEntry.getGroupId(),
+						objectDefinition.getExternalReferenceCode(),
+						objectEntry, _objectEntryService, objectField,
+						GuestOrUserUtil.getPermissionChecker(), _portal));
+				setMetadata(() -> _getMetadata(fileVersion, objectField));
+				setMimeType(dlFileEntry::getMimeType);
+				setName(dlFileEntry::getFileName);
+				setPreviewURL(
+					() -> _getPreviewURL(liferayFileEntry, objectField));
+				setScope(
+					() -> _getScope(
+						dlFileEntry, objectDefinition, objectEntry));
+				setThumbnailURL(
+					() -> _getThumbnailURL(dlFileEntry, objectField));
+			}
+		};
 	}
 
 	@Override
@@ -243,6 +339,16 @@ public class AttachmentObjectFieldBusinessType
 		}
 	}
 
+	private String _getFileBase64(
+			DLFileEntry dlFileEntry, ObjectField objectField)
+		throws Exception {
+
+		return NestedFieldsSupplier.supply(
+			objectField.getName() + ".fileBase64",
+			fieldName -> Base64.encode(
+				_file.getBytes(dlFileEntry.getContentStream())));
+	}
+
 	private Object _getFileEntryId(Object value) throws PortalException {
 		long fileEntryId = GetterUtil.getLong(value);
 
@@ -271,8 +377,174 @@ public class AttachmentObjectFieldBusinessType
 		return value;
 	}
 
+	private String _getFileURL(DLFileEntry dlFileEntry, ObjectField objectField)
+		throws Exception {
+
+		if (!Objects.equals(
+				ObjectFieldSettingConstants.VALUE_USER_COMPUTER,
+				ObjectFieldSettingUtil.getValue(
+					ObjectFieldSettingConstants.NAME_FILE_SOURCE,
+					objectField)) ||
+			GetterUtil.getBoolean(
+				ObjectFieldSettingUtil.getValue(
+					ObjectFieldSettingConstants.
+						NAME_SHOW_FILES_IN_DOCS_AND_MEDIA,
+					objectField.getObjectFieldSettings()))) {
+
+			return null;
+		}
+
+		return _exportImportAttachmentManager.getFileURL(dlFileEntry);
+	}
+
+	private Folder _getFolder(DLFileEntry dlFileEntry, ObjectField objectField)
+		throws Exception {
+
+		return NestedFieldsSupplier.supply(
+			objectField.getName() + ".folder",
+			fieldName -> {
+				if (!Objects.equals(
+						ObjectFieldSettingConstants.VALUE_DOCS_AND_MEDIA,
+						ObjectFieldSettingUtil.getValue(
+							ObjectFieldSettingConstants.NAME_FILE_SOURCE,
+							objectField))) {
+
+					return null;
+				}
+
+				return new Folder() {
+					{
+						setExternalReferenceCode(
+							() -> {
+								if (dlFileEntry.getFolderId() == 0) {
+									return null;
+								}
+
+								DLFolder dlFolder = dlFileEntry.getFolder();
+
+								return dlFolder.getExternalReferenceCode();
+							});
+						setSiteId(dlFileEntry::getGroupId);
+					}
+				};
+			});
+	}
+
+	private Map<String, Object> _getMetadata(
+			FileVersion fileVersion, ObjectField objectField)
+		throws Exception {
+
+		return NestedFieldsSupplier.supply(
+			objectField.getName() + ".metadata",
+			fieldName -> {
+				if ((fileVersion.getSize() == 0) ||
+					(!PDFProcessorUtil.hasImages(fileVersion) &&
+					 !PDFProcessorUtil.isDocumentSupported(
+						 fileVersion.getMimeType()))) {
+
+					return null;
+				}
+
+				return HashMapBuilder.<String, Object>put(
+					"numberOfPages",
+					PDFProcessorUtil.getPreviewFileCount(fileVersion)
+				).build();
+			});
+	}
+
+	private String _getPreviewURL(LiferayFileEntry liferayFileEntry)
+		throws PortalException {
+
+		if (StringUtil.startsWith(liferayFileEntry.getMimeType(), "video/")) {
+			return _dlURLHelper.getPreviewURL(
+				liferayFileEntry, liferayFileEntry.getFileVersion(), null,
+				"&videoEmbed=true", true, true);
+		}
+
+		return _dlURLHelper.getPreviewURL(
+			liferayFileEntry, liferayFileEntry.getFileVersion(), null,
+			StringPool.BLANK, false, false);
+	}
+
+	private String _getPreviewURL(
+			LiferayFileEntry liferayFileEntry, ObjectField objectField)
+		throws Exception {
+
+		return NestedFieldsSupplier.supply(
+			objectField.getName() + ".previewURL",
+			fieldName -> {
+				String previewURL = _getPreviewURL(liferayFileEntry);
+
+				if (Validator.isNull(previewURL)) {
+					return null;
+				}
+
+				return previewURL;
+			});
+	}
+
+	private Scope _getScope(
+			DLFileEntry dlFileEntry, ObjectDefinition objectDefinition,
+			ObjectEntry objectEntry)
+		throws Exception {
+
+		if ((objectEntry.getGroupId() == dlFileEntry.getGroupId()) &&
+			!Objects.equals(
+				objectDefinition.getScope(),
+				ObjectDefinitionConstants.SCOPE_COMPANY)) {
+
+			return null;
+		}
+
+		Group group = _groupLocalService.getGroup(dlFileEntry.getGroupId());
+
+		if (group.getType() == GroupConstants.TYPE_DEPOT) {
+			return Scope.ofReference(
+				group.getExternalReferenceCode(), Scope.Type.ASSET_LIBRARY);
+		}
+
+		return Scope.ofReference(
+			group.getExternalReferenceCode(), Scope.Type.SITE);
+	}
+
+	private String _getThumbnailURL(
+			DLFileEntry dlFileEntry, ObjectField objectField)
+		throws Exception {
+
+		return NestedFieldsSupplier.supply(
+			objectField.getName() + ".thumbnailURL",
+			fieldName -> {
+				String thumbnailURL = _dlURLHelper.getThumbnailSrc(
+					new LiferayFileEntry(dlFileEntry), null);
+
+				if (Validator.isNull(thumbnailURL)) {
+					return null;
+				}
+
+				return thumbnailURL;
+			});
+	}
+
 	private static final Log _log = LogFactoryUtil.getLog(
 		AttachmentObjectFieldBusinessType.class);
+
+	@Reference
+	private DLAppService _dlAppService;
+
+	@Reference
+	private DLFileEntryLocalService _dLFileEntryLocalService;
+
+	@Reference
+	private DLURLHelper _dlURLHelper;
+
+	@Reference
+	private ExportImportAttachmentManager _exportImportAttachmentManager;
+
+	@Reference
+	private File _file;
+
+	@Reference
+	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private Language _language;
@@ -281,6 +553,12 @@ public class AttachmentObjectFieldBusinessType
 	private ObjectDefinitionLocalService _objectDefinitionLocalService;
 
 	@Reference
+	private ObjectEntryService _objectEntryService;
+
+	@Reference
 	private ObjectScopeProviderRegistry _objectScopeProviderRegistry;
+
+	@Reference
+	private Portal _portal;
 
 }
